Add ability to text code from the references list
