+++ client/search-ui/src/components/FileMatchChildren.tsx
@@ -26,7 +26,7 @@ import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryServi
 import { codeCopiedEvent } from '@sourcegraph/shared/src/tracking/event-log-creators'
 import { useCodeIntelViewerUpdates } from '@sourcegraph/shared/src/util/useCodeIntelViewerUpdates'
 
-import { CodeExcerpt, FetchFileParameters } from './CodeExcerpt'
+import { CodeExcerpt, FetchFileParameters, onClickCodeExcerptHref } from './CodeExcerpt'
 import { LastSyncedIcon } from './LastSyncedIcon'
 
 import styles from './FileMatchChildren.module.scss'
@@ -40,53 +40,6 @@ interface FileMatchProps extends SettingsCascadeProps, TelemetryProps {
     fetchHighlightedFileLineRanges: (parameters: FetchFileParameters, force?: boolean) => Observable<string[][]>
     extensionsController?: Pick<ExtensionsController, 'extHostAPI'> | null
     hoverifier?: Hoverifier<HoverContext, HoverMerged, ActionItemAction>
-}
-
-/**
- * This helper function determines whether a mouse/click event was triggered as
- * a result of selecting text in search results.
- * There are at least to ways to do this:
- *
- * - Tracking `mouseup`, `mousemove` and `mousedown` events. The occurrence of
- * a `mousemove` event would indicate a text selection. However, users
- * might slightly move the mouse while clicking, and solutions that would
- * take this into account seem fragile.
- * - (implemented here) Inspect the Selection object returned by
- * `window.getSelection()`.
- *
- * CAVEAT: Chromium and Firefox (and maybe other browsers) behave
- * differently when a search result is clicked *after* text selection was
- * made:
- *
- * - Firefox will clear the selection before executing the click event
- * handler, i.e. the search result will be opened.
- * - Chrome will only clear the selection if the click happens *outside*
- * of the selected text (in which case the search result will be
- * opened). If the click happens inside the selected text the selection
- * will be cleared only *after* executing the click event handler.
- */
-function isTextSelectionEvent(event: MouseEvent<HTMLElement>): boolean {
-    const selection = window.getSelection()
-
-    // Text selections are always ranges. Should the type not be set, verify
-    // that the selection is not empty.
-    if (selection && (selection.type === 'Range' || selection.toString() !== '')) {
-        // Firefox specific: Because our code excerpts are implemented as tables,
-        // CTRL+click would select the table cell. Since users don't know that we
-        // use tables, the most likely wanted to open the search results in a new
-        // tab instead though.
-        if ((event.ctrlKey || event.metaKey) && selection.anchorNode?.nodeName === 'TR') {
-            // Ugly side effect: We don't want the table cell to be highlighted.
-            // The focus style that Firefox uses doesn't seem to be affected by
-            // CSS so instead we clear the selection.
-            selection.empty()
-            return false
-        }
-
-        return true
-    }
-
-    return false
 }
 
 /**
@@ -298,42 +251,15 @@ export const FileMatchChildren: React.FunctionComponent<React.PropsWithChildren<
     const viewerUpdates = useCodeIntelViewerUpdates(codeIntelViewerUpdatesProps)
 
     const history = useHistory()
-    /**
-     * This handler implements the logic to simulate the click/keyboard
-     * activation behavior of links, while also allowing the selection of text
-     * inside the element.
-     * Because a click event is dispatched in both cases (clicking the search
-     * result to open it as well as selecting text within it), we have to be
-     * able to distinguish between those two actions.
-     * If we detect a text selection action, we don't have to do anything.
-     *
-     * CAVEATS:
-     * - In Firefox, Shift+click will open the URL in a new tab instead of
-     * a window (unlike Chromium which seems to show the same behavior as with
-     * native links).
-     * - Firefox will insert \t\n in between table rows, causing the copied
-     * text to be different from what is in the file/search result.
-     */
     const navigateToFile = useCallback(
         (event: KeyboardEvent<HTMLElement> | MouseEvent<HTMLElement>): void => {
-            // Testing for text selection is only necessary for mouse/click
-            // events. Middle-click (event.button === 1) is already handled in the `onMouseUp` callback.
-            if (
-                (event.type === 'click' &&
-                    !isTextSelectionEvent(event as MouseEvent<HTMLElement>) &&
-                    (event as MouseEvent<HTMLElement>).button !== 1) ||
-                (event as KeyboardEvent<HTMLElement>).key === 'Enter'
-            ) {
-                const href = event.currentTarget.getAttribute('data-href')
-                if (!event.defaultPrevented && href) {
-                    event.preventDefault()
-                    if (props.openInNewTab || event.ctrlKey || event.metaKey || event.shiftKey) {
-                        openLinkInNewTab(href, event, 'primary')
-                    } else {
-                        history.push(href)
-                    }
+            onClickCodeExcerptHref(event, href => {
+                if (props.openInNewTab || event.ctrlKey || event.metaKey || event.shiftKey) {
+                    openLinkInNewTab(href, event, 'primary')
+                } else {
+                    history.push(href)
                 }
-            }
+            })
         },
         [props.openInNewTab, history]
     )
